requirements.txt
google-genai>=1.16.1
rich>=13.7.0
